chores: directory refactoring
