+++ src/main/java/com/unionclass/profileservice/domain/profile/dto/in/ChangeNicknameReqDto.java
@@ -1,6 +1,6 @@
-package com.unionclass.profileservice.profile.dto.in;
+package com.unionclass.profileservice.domain.profile.dto.in;
 
-import com.unionclass.profileservice.profile.vo.in.ChangeNicknameReqVo;
+import com.unionclass.profileservice.domain.profile.vo.in.ChangeNicknameReqVo;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
