+++ src/main/java/com/unionclass/profileservice/domain/profile/vo/in/RegisterNicknameReqVo.java
@@ -1,4 +1,4 @@
-package com.unionclass.profileservice.profile.vo.in;
+package com.unionclass.profileservice.domain.profile.vo.in;
 
 import lombok.Getter;
